initial commit
This is an example play to run my virt-infra Ansible role, which is found at:
https://github.com/csmart/ansible-role-virt-infra

Clone this repo recursively to pull down the role also, then run the
helper script to run against localhost. This will install required deps
and execute the sample inventory.

 $ git clone --recursive https://github.com/csmart/virt-infra-ansible
 $ cd virt-infra-ansible
 $ ./run.sh

Further instructions are included in the README.