shell commands, bump upstream 788624d
Upstream support was added to run shell commands on guest images as a
part of the disk-create task. This means you can do adhoc things, for
example, set the package mirror to use.

There is a new variable, simple add this into the inventory with a list
of shell commands you want to run. For example, setting a specific mirror
on Fedora:

 virt_infra_disk_cmd:
  - sed -i s/^metalink/#metalink/g /etc/yum.repos.d/*repo
  - sed -i s/^#baseurl/baseurl/g /etc/yum.repos.d/*repo
  - dnf config-manager --save --setopt baseurl=http://path-to-mirror fedora
  - dnf config-manager --save --setopt baseurl=http://path-to-mirror-updates updates

Or fastest mirror in Debian:

 virt_infra_disk_cmd:
   - apt-get update
   - apt-get install -y netselect-apt
   - netselect-apt -n -o /etc/apt/sources.list

See https://github.com/csmart/ansible-role-virt-infra/commit/788624df930bc49c5befc73829bad50188482588