update run script to work better from other dirs
The run script cd's into its directory so that the sample inventory and
ansible.cfg are found by default. This works great for testing, however
when executing the run script from another directory (such as an
inventory repo) the inventory won't be found if it's relative as the
script changes dir. Furthermore, you can't run a custom ansible.cfg.

So this simple change removes the cd call from the run script so that it
can be run from elsewhere. This doesn't change the default behaviour
when run from inside the directory, but it does mean that an inventory
source must be specified if run from elsewhere. However, this is
probably one of the main reasons to run it from another directory
anyway, so this seems reasonable.