ensure the role is present as git submodule
This example playbook is designed to use the Ansible role as a git
submodule.

However, if it was not cloned recursively, the run.sh will not do
anything, i.e.:

  $ ./run.sh
  BECOME password:
  PLAY [all] *****************************************************************
  PLAY RECAP *****************************************************************
  $

This change updates the README to include instructions on how to get the
clone (outside of TL;DR).

It also adds a check in the run.sh script which will automatically pull
the submodule if the directory does not exist.