bump role 2f17533, update for multiple kvmhosts
Bump to upstream role:
https://github.com/csmart/ansible-role-virt-infra/commit/2f17533699fec35b8d0dd3c4b8678fb3b9d0ed9e

This adds support for multiple KVM hosts (not just locahost), see the
commit above for details.

Add new inventory for simple_multikvm group which puts a VM on a
different KVM host. These are defined in kvmhost.yml inventory but point
to localhost. Update them to test on real KVM hosts.